Move common functionality to basic.py and DigitDecompExtend to decomp.py (#47)
* Refactor KeyMul and move it to basic.py
* Refactor and move DigitDecompExtend to its own file, decomp.py.
---------
